requirements.txt
anyio==4.11.0
appdirs==1.4.4
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
Jinja2==3.1.6
lseg-data==2.1.1
MarkupSafe==3.0.3
numpy==2.3.4
pandas==2.3.3
pyee==12.0.0
pyhumps==3.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.9.0
wheel==0.45.1
